requirements.txt
albumentations==1.4.3
datasets==2.12.0
gradio==4.31.4
keras==2.15.0
lightning==2.2.2
matplotlib==3.8.4
numpy==1.24.3
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pandas==2.2.2
Pillow==10.3.0
pycocotools==2.0.6
Shapely==2.0.4
SimpleITK==2.3.1
skimage==0.0
tensorflow==2.15.1
tensorflow_intel==2.15.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.2
transformers==4.37.2

